[FEATURE] GitCheckoutTask can checkout specific tag/sha1
In addition to checking out HEAD of a given branch the task can now
check out a specific tag or sha1.

The task can be given the options "sha1", "tag", "branch" now. They
are used in that order, i.e. sha1 has the highest precedence. If no
option is given, the default is branch "master".

Change-Id: I901506c747a13d083999ce42297e9ec9e09e3564