[!!!][TASK] Deployment and Workflow cleanups and tests
Nodes do not need to be registered both in an application and
deployment since the deployment will just use all nodes from every
registered application now. So the method addNode in Deployment was
removed and deployment configurations must be updated to remove that
line.

Additionally the initialization is checked in Workflow and Deployment
before running a deployment and some unit tests were introduced for the
most complex parts to ease future refactorings.

Change-Id: Ic659da6fc050f4affda8d91d43e3c3ba27b73f28