[BUGFIX] make SHA1 check in git checkout task more restrictive
Due to some weird SSH configuration, I had all SSH responses prefixed with
a fixed string; thus not only the SHA1 was returned but prefixed by some
string. This broke the remaining deployment, but in a way that it still
finished with "successful deployment".

This patch improves the error handling in this case.

Change-Id: Ic27dfca132f060354bb11102dc89bbb715562d11