[FEATURE] Better output with ANSI formatting
This change implements a custom log backend with ANSI formatting based
on tags. The deploy command has two new "verbose" and "no-ansi" switches
because the log threshold was updated to INFO and the command output
will be hidden by default for brevity.

In addition to the console output a file log backend will be used for
each deployment definition. It's also possible to override the default
logger through the deployment configuration with
$deployment->setLogger(...).

Change-Id: Ic07a5109b942d6cde115039bc3ee322d8bf1a1c0