[FEATURE] Argument for build configuration path
By default the deployment configuration are expected in "Build/Surf".
With the new argument "--configuration-path" this default path can be
overridden.

This change also introduces a new property for the deployment base path
in the Deployment model and adds a public method to the DeploymentService
for consistent access to deployment declarations and configuration.

Change-Id: I63dddc9547b8d94799356a2d5a0a05957cb79769