[FEATURE] Support flexible transfer and update methods
This change implements new stages and options for a declarative
configuration of the package / transfer / update process. The default is
to use no package method and a transfer method "git" that matches the
current behaviour. For Flow based application an update method "composer"
is the default for getting the current state of the application assets.

Additionally a new package method for a local git clone (with local
composer install) and a new transfer method "rsync" are implemented by
this change. An example configuration for this would be:

    $application->setOption('packageMethod', 'git');
    $application->setOption('transferMethod', 'rsync');
    $application->setOption('updateMethod', NULL);

The change should not be breaking for basic deployment configurations.

Change-Id: I92027f9e0365fd5c4dd2268d17b67dca52a81a7f