[FEATURE] Add recursiveSubmodules option to GitCheckoutTask
In some cases it is necessary to prevent git from checking out
submodules recursively during a Surf deployment in which composer is not
used. If the option "recursiveSubmodules" is set to false, only the
submodules directly referenced in the base repository are fetched and
checked out.

If the option is not set, or set to true, all submodules are fetched
and checked out recursively. Therefore, the default behaviour does not
change.

The unit tests adds a new possibility to match commands by a regular
expression to test for the absence of a --recursive flag.

Resolves: #49489
Change-Id: Idd1a934465c28767826559c9dbc8d73f96d9a0f9