[FEATURE] Add fetchAllTags option to GitCheckoutTask
Without further options, "git fetch" only fetches tags that point at
objects reachable from the branch heads (man 1 git-fetch). If a tag is
not fetched, the following step (git reset) will not proceed.
Therefore, it is necessary to force git to download all tags in some
circumstances.  This can be done by setting the option "fetchAllTags"
to true.

If the option is not set, or set to false, the default behaviour of
git is used, as was the case before. The default behaviour of Surf
therefore does not change.

Resolves: #50151
Change-Id: I2d17746bdcf786d10f034d758a75947e96ef191a