[BUGFIX] Copying config fails for subdirectories
When a file should be copied to a subdirectory of ``Configuration/*/``,
e.g. for a subcontext, the scp operation fails.

The (misleading) error message is "scp:
/var/www/.../Configuration/Production/Subcontext/: Is a directory" – the
real cause is that the directory does not exist and thus scp would
rather copy the file to the folder path, which is pointless as the path
points to a directory (as the path ends in a slash).

Solution is to execute a ``mkdir -p ...$configurationPath`` before
copying the file.

Change-Id: If00cb0dd133ed24a511224c7092e4938ec5fc9fb
Resolves: #48652