[!!!][FEATURE] Make symlinkconfiguration use application context
This change does the following regarding the context that is symlinked:

* uses the context of the application if it is a Flow application and
  no explicit context has been set
* falls back to "Production"

This might be breaking in rare cases, but should be working unchanged
for most people. Use ``surf:simulate --verbose`` and check the steps
taken to make sure it works for your case.

Additionally the BaseTaskTest is extended to support an array of
commands and the AssertCommandExecuted constraint is fixed for recent
PHPUnit versions.

Change-Id: Ib5421d585c1870e655ab96b4cbfe71050be4278e