[BUGFIX] Fix creating shared/Data directories in Flow application
With commit 642f7c3 the base directory for creating these directories
changed from the deployment base to the release path, causing the
directories to be created in the wrong place.

This commit adds a new option to allow specifing a different base
directory and the flow task can then provide this with the deployment
base.

Related: #25