[!!!] Remove applicationWebDirectory and applicationRootDirectory
The options have been removed. In many places Surf assumes that the
application root is exactly the git checkout root. Making this
configurable would mean a major refactoring with no benefits for the
default case.

The web directory however stays configurable with the `webDirectory`
option. This must be set to a path relative to the checkout root
and defaults to `web` for the TYPO3/CMS application.

Fixes #38