[!!!] Make Surf work with latest TYPO3 Console versions
TYPO3 Console was converted into a composer library, thus
we must not expect it to be located in typo3conf/ext

Additionally the typo3cms binary is now located in the vendor/bin
directory by default, thus we make this a default in Surf as well.

In case you use older TYPO3 Console Versions or not a composer installation,
you can configure a different path like that:
`$application->setOption('scriptFileName', './typo3cms');`

The path needs to be relative to the git checkout root.

Fixes #50